Connect_Four_Game
Connect Four is a two-player connection game firstly insert your name,
in which players first choose a color and then take turns dropping colored disfrom the top into a seven-column,
six-row vertically suspended grid. The pieces fall straight down, occupying the next available space within the column.
The objective of the game is to be the first to form a horizontal, vertical or diagonal line of four of one's own discs.
Then connect Four is a solved game.